requirements.txt
crewai>=0.55
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.8
python-dotenv>=1.0
streamlit>=1.49
google-genai>=0.3   # Gemini 2.5 Flash/Pro API client (new SDK)
